# Record expr location for opqaue value numbers