… src/jit/valuenum.cpp