@JosephTremoulet 2016-08-16
@dotnet/jit-contrib, PTAL.  This allows loop-invariant code hoisting to kick in in 11 of the benchmarks under the JIT/Performance tests, 11 of the core framework libraries (a handful of methods in each), 84 methods in desktop mscorlib, ~250 methods in desktop frameworks, and 80-100 methods in most of the other SuperPMI asm diff inputs.  Also gets each of the individual cases discussed in #6303.