@russellhadley 2016-08-16
@JosephTremoulet Do we have perf numbers for this yet?