@JosephTremoulet 2016-08-16
@russellhadley, did some local runs but changes seemed to be in the noise range.  Here's the raw data:

Benchmark | Diffs? | Delta (negative => improved, positive => regressed) | Avg(ms) (master) | Avg(ms) (VNLoops) | Min(ms) (master) | Max(ms) (master) | StdDev(%) (master) | Min(ms) (VNLoops) | Max(ms) (VNLoops) | StdDev(%) (VNLoops)
------------------|----|---------|------|------|-----|------|-----|------|-----|-----
BenchI-8Queens | No | -0.63% | 158 | 157 | 158 | 159 | 0.57 | 157 | 158 | 0.4
BenchI-Ackermann | No | -1.26% | 159 | 157 | 158 | 163 | 1.26 | 157 | 159 | 0.64
BenchI-AddArray | Yes | 0.00% | 157 | 157 | 157 | 158 | 0.57 | 157 | 159 | 0.64
BenchI-AddArray2 | No | -1.73% | 173 | 170 | 172 | 175 | 0.63 | 169 | 172 | 0.64
BenchI-Array1 | No | -0.63% | 160 | 159 | 159 | 161 | 0.56 | 157 | 161 | 0.89
BenchI-Array2 | No | -3.07% | 163 | 158 | 163 | 165 | 0.67 | 158 | 159 | 0.28
BenchI-BenchE | No | 0.00% | 158 | 158 | 157 | 161 | 1.06 | 158 | 159 | 0.28
BenchI-BubbleSort | No | -1.26% | 159 | 157 | 159 | 160 | 0.4 | 157 | 158 | 0.49
BenchI-BubbleSort2 | Yes | -1.26% | 159 | 157 | 158 | 161 | 0.84 | 156 | 159 | 0.75
BenchI-CSieve | Yes | 0.00% | 157 | 157 | 156 | 158 | 0.49 | 157 | 158 | 0.4
BenchI-Fib | No | 0.00% | 157 | 157 | 157 | 158 | 0.4 | 157 | 158 | 0.4
BenchI-HeapSort | Yes | -0.63% | 159 | 158 | 158 | 161 | 0.69 | 157 | 159 | 0.57
BenchI-IniArray | No | 0.00% | 157 | 157 | 156 | 158 | 0.57 | 157 | 158 | 0.4
BenchI-LogicArray | No | -0.63% | 159 | 158 | 159 | 161 | 0.69 | 157 | 160 | 0.69
BenchI-Midpoint | Yes | -0.63% | 159 | 158 | 159 | 160 | 0.28 | 158 | 160 | 0.69
BenchI-MulMatrix | No | -5.85% | 171 | 161 | 170 | 173 | 0.74 | 161 | 163 | 0.68
BenchI-NDhrystone | No | -4.82% | 166 | 158 | 165 | 167 | 0.54 | 158 | 159 | 0.49
BenchI-Permutate | No | -0.63% | 159 | 158 | 159 | 161 | 0.56 | 158 | 158 | 0
BenchI-Pi | No | 0.00% | 164 | 164 | 163 | 165 | 0.47 | 164 | 166 | 0.67
BenchI-Puzzle | No | -7.02% | 171 | 159 | 171 | 173 | 0.52 | 158 | 160 | 0.4
BenchI-QuickSort | Yes | 0.00% | 158 | 158 | 158 | 160 | 0.69 | 157 | 159 | 0.4
BenchI-TreeSort | No | -1.24% | 161 | 159 | 161 | 162 | 0.48 | 158 | 160 | 0.4
BenchI-TreeInsert | No | -0.63% | 159 | 158 | 159 | 160 | 0.56 | 158 | 158 | 0
BenchI-XposMatrix | No | -0.63% | 159 | 158 | 159 | 160 | 0.4 | 157 | 159 | 0.49
BenchF-Adams | No | -10.67% | 253 | 226 | 249 | 264 | 2.16 | 223 | 235 | 1.94
BenchF-BenchMk2 | No | -2.44% | 164 | 160 | 159 | 173 | 3.23 | 160 | 161 | 0.28
BenchF-BenchMrk | No | 0.00% | 160 | 160 | 159 | 161 | 0.56 | 160 | 160 | 0
BenchF-Bisect | No | -4.85% | 165 | 157 | 165 | 167 | 0.61 | 157 | 159 | 0.7
BenchF-DMath | No | -1.25% | 160 | 158 | 160 | 161 | 0.4 | 157 | 160 | 0.69
BenchF-FFT | No | -1.23% | 162 | 160 | 161 | 163 | 0.62 | 159 | 162 | 0.68
BenchF-InProd | No | -2.45% | 163 | 159 | 163 | 165 | 0.67 | 158 | 160 | 0.4
BenchF-InvMt | Yes | -2.47% | 162 | 158 | 161 | 163 | 0.48 | 158 | 160 | 0.63
BenchF-LLoops | Yes | -20.69% | 203 | 161 | 202 | 204 | 0.44 | 161 | 162 | 0.48
BenchF-Lorenz | No | -1.26% | 159 | 157 | 158 | 161 | 0.89 | 157 | 158 | 0.49
BenchF-MatInv4 | Yes | -7.65% | 170 | 157 | 169 | 173 | 0.98 | 156 | 158 | 0.57
BenchF-NewtE | No | -4.22% | 166 | 159 | 166 | 168 | 0.54 | 159 | 160 | 0.4
BenchF-NewtR | No | -4.27% | 164 | 157 | 164 | 165 | 0.47 | 157 | 159 | 0.64
BenchF-Regula | No | -4.85% | 165 | 157 | 164 | 166 | 0.47 | 157 | 158 | 0.57
BenchF-Romber | Yes | -4.17% | 168 | 161 | 168 | 170 | 0.6 | 160 | 162 | 0.39
BenchF-Secant | No | -5.42% | 166 | 157 | 164 | 169 | 1.08 | 157 | 158 | 0.49
BenchF-Simpsn | No | -2.44% | 164 | 160 | 164 | 165 | 0.47 | 159 | 162 | 0.68
BenchF-SqMtx | Yes | -0.63% | 158 | 157 | 158 | 159 | 0.57 | 157 | 158 | 0.4
BenchF-Trap | No | -3.64% | 165 | 159 | 164 | 167 | 0.66 | 159 | 160 | 0.49
BenchF-Whetsto | No | -3.61% | 166 | 160 | 165 | 167 | 0.47 | 159 | 162 | 0.79
binarytrees | No | -0.52% | 2878 | 2863 | 2866 | 2897 | 0.42 | 2857 | 2876 | 0.24
fasta | No | 4.91% | 2608 | 2736 | 2559 | 2699 | 2.26 | 2698 | 2809 | 1.41
fastaredux | No | 1.83% | 3504 | 3568 | 3431 | 3607 | 1.66 | 3484 | 3614 | 1.28
nbody | No | -1.34% | 2309 | 2278 | 2299 | 2344 | 0.76 | 2264 | 2308 | 0.75
pidigits | No | -7.13% | 659 | 612 | 658 | 661 | 0.18 | 608 | 616 | 0.5
spectralnorm | No | -0.13% | 3746 | 3741 | 3745 | 3748 | 0.04 | 3740 | 3743 | 0.04

The 20% win in lloops did not reproduce on re-run.