@russellhadley 2016-08-16
@JosephTremoulet There are notes about the code we expect to get rid of for BenchI/Bubblesort2 and BenchI/Sq_mtx  (though I'm not seeing that one in the list - maybe they meant BenchF/SqMtx?)  Are you seeing that code being optimized the way we expect it to be?