@JosephTremoulet 2016-08-17
@russellhadley took a look at SqMtx -- we're hoisting everything we want to hoist, though a couple redundant bounds-checks remain in the loop due to #6779.