@russellhadley 2016-08-17
Thanks for running down the extra details and opening the item about assertion prop.