@JosephTremoulet 2016-08-18
We're not getting the hoist in BubbleSort2 due to #6790.